requirements.txt
# requirements.txt recomendado para DeepBDE Backend
Django>=4.2
djangorestframework
rdkit
torch
numpy
pandas
pillow
dgl
scikit-learn
django-tex
PyJWT
fpdf2
drf-spectacular
networkx==2.8.8
